Move initialization logic to instance
